{"types":"C","runs":3}

+++ Problem16.java
@@ -1,5 +1,4 @@
 import acm.program.ConsoleProgram;
-import java.lang.*;
 
 public class Problem16 extends ConsoleProgram{
 	
@@ -7,7 +6,7 @@ public class Problem16 extends ConsoleProgram{
 		int a = readInt("Enter a: ");
 		int b = readInt("Enter b: ");
 		int c = readInt("Enter c: ");
-		double harmonicaverage = 1.0 / (a+b+c);
+		double harmonicaverage = (double)(1 / a + b + c );
 		println("Harmonic average is " + harmonicaverage + ".");
 	}
